add master api
