+++ protocols/BUILD
@@ -9,6 +9,17 @@ proto_library(
     deps = [":pain_common_proto"],
 )
 
+proto_library(
+    name = "pain_errno_proto",
+    srcs = ["errno.proto"],
+)
+
+cc_proto_library(
+    name = "cc_pain_errno_proto",
+    deps = [":pain_errno_proto"],
+    visibility = ["//visibility:public"],
+)
+
 cc_proto_library(
     name = "cc_pain_common_proto",
     deps = [":pain_common_proto"],
@@ -17,6 +28,8 @@ cc_proto_library(
 
 cc_proto_library(
     name = "cc_pain_master_proto",
-    deps = [":pain_master_proto"],
+    deps = [
+        ":pain_master_proto",
+    ],
     visibility = ["//visibility:public"],
 )
