+++ protocols/common.proto
@@ -2,9 +2,16 @@ syntax = "proto3";
 
 package pain.proto;
 
-enum ChunkFlags {
-    CHUNK_FLAG_NONE = 0;
-    CHUNK_FLAG_SEALED = 1;
+message Header {
+    uint32 status = 1;
+    string message = 2;
+}
+
+enum ChunkState {
+    CHUNK_STATE_NONE = 0;
+    CHUNK_STATE_INIT = 1;
+    CHUNK_STATE_CHECKIN = 2;
+    CHUNK_STATE_SEALED = 3;
 }
 
 message Location {
@@ -15,17 +22,14 @@ message Location {
 message SubChunkInfo {
     string sub_chunk_id = 1;
     uint64 sub_chunk_size = 2;
-    uint64 sub_chunk_offset = 3;
-    ChunkFlags flags = 4;
-    Location location = 5;
+    Location location = 3;
 }
 
 message ChunkInfo {
     string chunk_id = 1;
     uint64 chunk_size = 2;
-    uint64 chunk_offset = 3;
-    ChunkFlags flags = 4;
-    repeated SubChunkInfo sub_chunks = 5;
+    ChunkState state = 3;
+    repeated SubChunkInfo sub_chunks = 4;
 }
 
 message FileInfo {
